[Refactor] : env file 형식 yaml으로 변경
Git : #21
