+++ .gitignore
@@ -3,8 +3,6 @@
 __pycache__
 .pytest_cache
 
-eb_fast_api/env/.env
-
-jenkins/
-
 *.db
+
+eb_fast_api/env/sources/env.yaml
