+++ develop.Dockerfile
@@ -5,5 +5,4 @@ COPY . .
 RUN pip install -e .
 RUN pip install -r requirements.txt
 
-WORKDIR /eb_fast_api
-ENTRYPOINT ["uvicorn", "main:app", "--host", "0.0.0.0", "--port", "8001"]
+ENTRYPOINT ["sh", "./script/develop-script.sh", "8001"]
