+++ requirements.txt
@@ -31,6 +31,7 @@ pytest-asyncio==0.23.8
 python-dotenv==1.0.1
 python-jose==3.3.0
 python-multipart==0.0.9
+PyYAML==6.0.2
 rsa==4.9
 setuptools==68.2.2
 six==1.16.0
